Pull cffdrs from artifactory
