+++ api/pyproject.toml
@@ -44,6 +44,8 @@ httpx = "^0.27.0"
 orjson = "^3.9.0"
 greenlet = "^3.0.0"
 sentry-sdk = {extras = ["fastapi"], version = "^2.0.1"}
+cffdrs = {version = "^0.1.0", source = "psu"}
+numba = "^0.59.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.1"
@@ -61,6 +63,12 @@ jsonpickle = "^3.0.0"
 pytest-watch = "^4.2.0"
 pytest-testmon = "^2.0.0"
 ruff = "^0.4.0"
+
+
+[[tool.poetry.source]]
+name = "psu"
+url = "https://artifacts.developer.gov.bc.ca/artifactory/api/pypi/pe1e-gen-python-local/simple"
+priority = "supplemental"
 
 [build-system]
 requires = ["poetry>=1.1.11"]
